Fixed bug in plugins/ccxinp.py to read planar elements.
Planar element are missing the z coordinates so the reader was wrongly treating the nodes,
as it was trying to read 4 elements at the time, while for planar stress elements only 3 numbers
per node line are defined.
Now a check on the first line of the node data is performed to correctly format the data.