Add a fix for to avoid a bug in fewgl
The standard display module to render pyFormex webgl models
contains a bug in the reading of pgf format files.

A workaround has been added in the export of PGF format files
to avoid triggering that bug. The fix may change a single coordinate
value with the smallest representable amount on the computer
and is thus harmless. It is therefore activated by default,
but can be disabled by setting the configuration variable
webgl/avoid_fewgl_read_pgf_bug to False.