Replace optparse module with argparse
The parsing of the command line parameters is now done by the
newer argparse module, instead of the deprecated optparse.
Some of the options have been slightly changed in order to make
use of newer features in argparse.
The --testmodule option has been renamed to --doctest. This is
because we want to introduce a second testing framework besides
doctest: pytest allows more complex testing than what can be done
with docstrings.

The command to run the (docstring) tests for modules coords and curve
was before:
  pyformex --testmodule coords,plugins.curve
Now this becomes:
  pyformex --doctest coords plugins.curve
and when pytest tests will be written for the modules:
  pyformex --pytest coords plugins.curve

See pyformex --help for all options.