Print Attributes with keys sorted.
The Attributes dict is now printed with its keys sorted.
This was needed to pass the doctests, because Python3 has
the dict entries in another (and not constant) order than Python2.
Maybe we will turn this __str__ function into a general dict
printing function, but it obviously only works for dicts where
the keys can be sorted (e.g. all strings, which is the case for
Attributes).