Fix doctests in olist for Python3
In Python3 range does not return a list, but a range, so the
examples have been changed to use 'lrange' instead of range.

lrange is a new function in olist which returns a range as a
list instead. Takes same parameters as range().
lrange should be used wherever you need the list of all numbers
in the range. For looping over a range however, you should still
use the range function.