Force absolute imports in pyFormex
All imports in pyFormex were already converted to absolute imports,
i.e. starting from the pyformex pkg. Thus:
   from pyformex import coords
or
   from . import coords    (in a module in same package)
but not just
   import coords
(which is allowed in Python2, but not in Python3).
Now we are making this notation obligatory, by adding a
  from __future__ import absolute_import
to all modules. If any deprecated import is still lingering
somewhere, it will probably show up now.