Convert arraytools module to Python3 division
Python3 integer division is different from Python2.
In Python2 integer division (/) truncates to an integer.
In Python3 integer division results in a float. Truncating
division has to be written as //.
This means that in order to make the same pyFormex code working
under both PY2 and PY3, we need to check the source for every
division operator (/ and /=), check the types of the operands,
and if both are integer, ask ourselves whether the intention was
to truncate and have an integer, or not.
If the intention was indeed truncation, we need to write //.
The process can be done file by file. After everything has been
converted in a file, the Python3 division type should be
activated by adding 'division' in the from __future__ import
statement at the top of the module.

In most cases where truncation was not wanted, we explicitely
converted one of the arguments to float. While these conversions
could be left away in future, we will not bother doing so.