Add some PY2/PY3 compatibility functions
- pyformex.round(number,ndigits): performs Python3 rounding behavior
Values ending on 5 are rounded to the nearest even number.
It is recommended to add 'from pyformex import round' in all code
where you use the builtin round function. This will result in the
same behavior under py2 and py3. Note that
   round(number)  returns an int
   round(number, 0)  returns a float with 0 fraction
   round(number, ndigits) returns a float rounded to ndigits
Note that numpy round/around has the same behavior, except for
the int return when ndigits is omitted.

- pyformex.isFile and pyformex.isString can be used to test whether
an object is a file, or a string (ascii/bytes or unicode) type.