Remove some website decorations from the internal docs.
The W3C validator, FSF plugger, and Paypal donate button are removed
from the internal docs. They will only remain visible on the
documentation that displayed on the pyFormex website.

These decorations indeed have no direct tie to the software itself
and its use. And they are against Debian regulations for docs, so
need to be stripped off anyways when creating .deb packages.

It is easier then to strip them off always, and only put them in
when creating the website docs.