Allow swapped numbering direction in simple.regularGrid
The nodes in a multidimensional grid created by simple.regularGrid
are numbered first in the direction of the highest axis, the the
next highest and so down to the lowest axis. This is contrary to
the default numbering scheme used in most other pyFormex functions
and classes, which number first in the direction of the lowest (0)
axis.

Now an option swapaxes is added to the regularGrid function.
If True, the noncompliant order of the old behavior is retained.
If False, the numbering scheme will be in accordance with
default pyFormex behavior.

In this commit, the default value is set to True, for compatibility
reasons: old code should work as before, but you are encouraged
to check your code and add 'swapaxes=True' if you depend on the old
numbering order.

In a next commit, the default will be made to False, and a warning
will be added if you do not use the 'swapaxes' argument.