Add parameter to pf.GUI.fullScreen; change shortcut key to F11
The pf.GUI.fullScreen function now has an optional bool parameter
to force the display to full screen (True) or normal window mode
(False). The default is to toggle the mode.
The shortcut key to toggle the mode is changed from F5 to F11, as
that seems to be the mostly used one nowadays.
(Where are the good old F5 days? Was that WordPerfect?)