Add a busy() function to pyFormex scripting language.
Use busy(True)...busy(False) around any part that will/might
take a very long time to execute.
In a GUI execution, this will change the cursor shape
(equivalent with pf.GUI.setBusy()),
while in a non-GUI, it will write a message to be patient.

Because it works in both modes, ans is easier than pf.GUI.setBusy(),
the latter should be replaced everywhere with busy().