Allow inline script in the pyformex command line
The pyFormex command now has an option --script SCRIPT to execute
a short inlined script in pyFormex. It is mostly useful together
with the --nogui option. There is also a short option version: -c.
As an example, the command
  pyformex --nogui -c "import sys;print(sys.path)"
will print the Python sys.path. And
  pyformex --nogui -c "import vtk;print(vtk)"
will show from where vtk is imported.
Maybe we should make --nogui default when using this option?