Add --experimental option to pyformex command
The --experimental option adds the directory
PYFORMEX_INSTALL_PATH/pyformex/experimental to the sys.path, at position 1
(behind the PYFORMEX_INSTALL_PATH).
This will allow to drop modules in there that should override modules
normally loaded from other paths.
This is only intended for experiments, not for production.
The experimental directory will not be released (though it might be
included in the repository).