Modularization of pyFormex startup code.
This is a further step in the modularization of the pyFormex
startup code. This will enable easier import of pyFormex in a
standard Python workflow.

The default configuration (from pyformex/pyformexrc) is now
loaded automatically during initialization of pyFormex (which
is done during import of the pyformex package, which is always
the first import (explicit or implicit).

Thus, setting the Python path correctly and doing
  from pyformex.script import *
suffices to have the full pyFormex script language
available in Python. See the revised examples in the
importing_pyformex.rst document.