Added parameter check_surface in vtk_itf.pointInsideSurface
this is a  boolean flag to first check whether the surface is
a closed manifold. If True and the surface is not a closed manifold, all points
will be marked outside. Note that if this check is not performed and the surface
is not closed, the results are undefined. the function was modified accordingly.