Fix a bug in CoordSys, change meaning of rot attribute
The rotation matrix in CoordSys is now stored in the same way
as arraytools rotationMatrix and rotmat functions.
This means that the dot product of the global axes unit vectors
and the rot matrix returns the new axes unit vectors.
The latter are precisely the ROWS of the rot matrix (before
it was the COLUMNS).
This also fixes a bug in the oab initialization of the CoordSys,
where a transpose was forgotten, and simplifies the code.
It also fixes a bug in the CoordSys.rotate method.

The change also provides more consistency and ease of use, since
rotation matrices from different sources are now stored in the same
way (The inertia module still uses a transposed storage scheme, so
we might change that one also).

The change has no influence on the meaning of Coords.toCS and fromCS:
these methods have been adapted to the new storage scheme.