Improved FieldColor example
The FieldColor example now also demonstrates how to interactively
change colors on a Mesh type geometry (with nplex > 1).
Note that in most cases you want to change both front and back color
(drawable 0 and 1).

The opengl.drawable.changeVertexColor now tests that the drawable uses
a vertex color buffer, and that the size of the provided color array
matches.