Change CoordSys translate and rotate methods
CoordSys.rotate and translate now operate exactly as the Coords
methods. They return a new CoordSys obtained from applying the
transformation on the CoordSys.points() (thus rotation is in the
global axes).
Two low level methods: _translation and _rotation apply a simple
translation vector or rotation matrix to the CoordSys itself,
changing the object inplace. Rotation in this case is around the
origin of the CoordSys.