Add comment option to AbqData.write
The comment option has been added to deal with most of the issues
of the copy_script option.
If a comment is specified, it will be included as comments in the
INP file, at the proper place.
The fmtHeading function was therefore split in three parts:
- fmtWatermark() provides the start lines of the file
- fmtComment() formats the comments
- fmtHeading() formats the *HEADING command

This is only implemented in fe_abq_new, as the old interface will
soon be removed.