Allow RGBA color mode in draw function.
pyFormex can now draw using colors specified with four components
(RGBA). This allows for varying alpha values over the structures.

Currently the feature is experimental, and you need adapted
shader programs. Use the --shader=alpha on the pyformex command line.

There is an example showing the effect of varying alpha values:
ColorRGBA.

Setting 4-component color as attrib on an object allows you to
interactively change all four components, thus also the transparency.
There is no example for this yet.

This new feature is currently only activated by a command line option,
because the new alpha shader forces all colors to be 4-component,
which is a waste of precious GPU memory if you do not want to use
varying alpha. Still we encourage users to experiment with the
new shader.