Display a warning from WebGL export about using soft-zooming.
Background: pyFormex has three ways of zooming:
- dolly: moving the camera to/from the scene
- fovy: change the lens angle
- soft: change the area of the lens view mapped to the viewport (this is
  like software zooming on a physical camera).
The currently used WebGl displayer (fewgl.js) however only has dolly
zooming: you have to set the camera at the correct position.
When creating a WebGL export from a pyFormex scene that has soft zooming
(only a part of the camera image is mapped to the viewport: this can be
seen from pf.canvas.camera.report() : Area is not (0.,0.), (1.,1.)),
the exported WebGL image may be different from the pyFormex view,
because the fewgl.js currently does not have the soft-zooming feature.
Thus, if you export from a scene with soft-zooming, the exported image
will be different in size.

Therefore, a warning is now displayed from the WebGL exporter when you
are exporting a soft-zoomed scene. The warning also contains a hint on
how to avoid that situation.