Increase functionality of PolyLine.close and PolyLine.open
PolyLine.close until now always added an extra segment to the
PolyLine, even if the first and last pointof the (open) PolyLine
coincided. Now there is an option to specify a tolerance, and if
the end points of the open curve are closer than this value,
no new segment is added, and the last point is remove from the
curve instead. The close method returns a value True if the
curve was closed without adding a segment. This may be helpful
in properly re-opening the curve.

The inverse operation PolyLine.open until now always removed the
last segment of the curve. In order to mirror the new close
functionality, open() now takes an optional argument keep_last.
If set True, the PolyLine will be opened by adding a point and
keeping all segments.